fix: dont delete if simprints is in the list
Former-commit-id: 2be7596057293d391f3fa030f6356624686fcbb0 [formerly 00ae1d3d24835229dafcfa9a19cd6b1699e2c7ef]
Former-commit-id: 67dba30fcbe19c2fb7a25f03c736f9f2f68540b5

+++ clientapi/src/main/java/com/simprints/clientapi/activities/commcare/CommCarePresenter.kt
@@ -174,8 +174,12 @@ class CommCarePresenter(
         processConfirmIdentityRequest()
     }
 
+    /**
+     * Delete the events if returning to CommCare but not Simprints
+     */
     private suspend fun deleteSessionEventsIfNeeded(sessionId: String) {
-        if (sharedPreferencesManager.syncDestinationSetting.contains(SyncDestinationSetting.COMMCARE)) {
+        if (sharedPreferencesManager.syncDestinationSetting.contains(SyncDestinationSetting.COMMCARE) &&
+            !sharedPreferencesManager.syncDestinationSetting.contains(SyncDestinationSetting.SIMPRINTS)) {
             sessionEventsManager.deleteSessionEvents(sessionId)
         }
     }
